feat(cli,daemon): show daemon+cli version in status / stats / system-status
Adds a one-line version display to the three observability commands so
a stale long-running daemon paired with a freshly-upgraded CLI binary
(or the reverse) is visible at a glance instead of inferred from the
existing mtime-based stale-binary heuristic.

Wire-protocol change:
  StatusResponse and StatsResponse each get a new `version: String`
  field carrying the daemon binary's compile-time `env!("CARGO_PKG_VERSION")`.
  Both fields are `#[serde(default)]` so a new CLI talking to a
  pre-this-commit daemon decodes a missing field as an empty string
  and renders it as `<unknown> (daemon) / <cli> (cli)` rather than
  failing the deserialize.  Same back-compat as every other recent
  protocol field (mimalloc_committed_bytes, agg_cache_*, etc.).

CLI rendering:
  New `crate::commands::version_summary(&daemon_version)` helper
  returns one of three shapes based on the comparison to the CLI's
  own `env!("CARGO_PKG_VERSION")`:

    daemon == cli              "0.5.78"
    daemon != cli              "0.5.77 (daemon) / 0.5.78 (cli)  ⚠ MISMATCH"
    daemon empty (legacy)      "<unknown> (daemon) / 0.5.78 (cli)"

  Used by:
    daemon_mgmt::daemon_status     `Version:       <value>`
    daemon_mgmt::daemon_stats      `Version:           <value>`
    system_status::print_daemon_status   `  Version:     <value>`

  Indentation/padding matches each command's existing label column.
  The mtime-based stale-binary warning in system_status stays in
  place — the two signals are complementary: mtime catches "daemon
  started before last build", version-mismatch catches "client and
  daemon are different builds".

Daemon population:
  index::IndexManager::status() and stats() each set
  `version: env!("CARGO_PKG_VERSION").to_owned()`.  Same source of
  truth the existing `DaemonEvent::DaemonStarting` event already uses
  in lib.rs:284, so client-side observers see consistent strings
  regardless of whether they tap the event stream or the RPC.

File-size policy:
  Adding the `version` field + its doc comment pushes
  crates/uffs-client/src/protocol/response.rs from 800 to 826 LOC
  (26 over the policy cap).  Registered as a PERMANENT exception in
  scripts/ci/file_size_exceptions.txt with a matching `//! Exception:`
  block in the file: this is the single declaration site for every
  wire-protocol Params/Response struct, splitting would fragment the
  cohesive contract across multiple files for every protocol consumer
  to chase.

Validation:
  * cargo check -p {client,daemon,cli} --tests              clean
  * RUSTDOCFLAGS=-Dwarnings cargo doc -p {…} --no-deps       clean
  * cargo nextest run -p {client,daemon,cli} --lib          271 / 271
  * bash scripts/ci/check_file_size_policy.sh                ALLOW
  * Live release-binary smoke against the user's currently-running
    pre-this-commit daemon prints
      `Version:       <unknown> (daemon) / 0.5.78 (cli)`
    confirming the back-compat path.
